Deployment/requirements.txt
tensorflow==2.17.1
keras==3.5.0
streamlit==1.25.0
pillow<10.0.0,>=7.1.0
opencv-python-headless==4.8.0.74
numpy==1.23.5
matplotlib==3.9.4
seaborn==0.13.2